# docu/rest/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.2",
        "title": "Chimera API"
    },
    "host": "localhost:8080",
    "basePath": "/api/interface/v2",
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/scenario/{scenarioID}/instance/{instanceID}/activity": {
            "get": {
                "description": "Returns a JSON-Object containing information about all activity\n instances of a specified scenario instance.",
                "operationId": "getActivitiesOfInstance_",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "scenarioID",
                        "in": "path",
                        "required": true,
                        "x-is-map": false,
                        "type": "integer",
                        "format": "int64",
                        "description": "The id of the scenario"
                    },
                    {
                        "name": "instanceID",
                        "in": "path",
                        "required": true,
                        "x-is-map": false,
                        "type": "integer",
                        "format": "int64",
                        "description": "The id of the instance."
                    },
                    {
                        "name": "filter",
                        "in": "query",
                        "required": false,
                        "x-is-map": false,
                        "type": "string",
                        "description": "Defines a search strings. Only activities\n                     with a label containing this String will be\n                     shown."
                    },
                    {
                        "name": "state",
                        "in": "query",
                        "required": false,
                        "x-is-map": false,
                        "type": "string",
                        "description": "The state of the instance."
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list with all activity instances for the scenario instance.",
                        "schema": {
                            "$ref": "#/definitions/Activities"
                        }
                    },
                    "404": {
                        "description": "The scenario or instance wasn't found.",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
...
        "/scenario/{scenarioID}/instance/{instanceID}/activityinstance/{activityInstanceId}/workingItems": {
            "get": {
                "description": "This method provides a list of working items for an activity instance.",
                "operationId": "getWorkingItems_",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "scenarioID",
                        "in": "path",
                        "required": true,
                        "x-is-map": false,
                        "type": "integer",
                        "format": "int64",
                        "description": "The databaseID of a scenario."
                    },
                    {
                        "name": "instanceID",
                        "in": "path",
                        "required": true,
                        "x-is-map": false,
                        "type": "integer",
                        "format": "int64",
                        "description": "The databaseID of a scenarioInstance."
                    },
                    {
                        "name": "activityInstanceId",
                        "in": "path",
                        "required": true,
                        "x-is-map": false,
                        "type": "integer",
                        "format": "int64",
                        "description": "The databaseID of an activityInstance."
                    }
                ],
                "responses": {
                    "202": {
                        "description": "A list of working items",
                        "schema": {
                            "$ref": "#/definitions/WorkingItems"
                        }
                    },
                    "404": {
                        "description": "The scenario/scenarioInstance/activityInstance was not found.",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Activities": {
            "type": "object",
            "properties": {
                "activities": {
                    "$ref": "#/definitions/ActivityList"
                },
                "ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    },
                    "description": "List of all activity ids"
                }
            }
        },
        "ActivityList": {
            "type": "object",
            "description": "Details for all activites",
            "properties": {
                "(id)": {
                    "type": "object",
                    "properties": {
                        "activityid": {
                            "type": "number",
                            "format": "integer"
                        },
                        "link": {
                            "type": "string"
                        },
                        "id": {
                            "type": "number",
                            "format": "integer"
                        },
                        "label": {
                            "type": "string"
                        },
                        "state": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "WorkingItems": {
            "type": "array",
            "description": "List of working items",
            "items": {
                "$ref": "#/definitions/WorkingItem"
            }
        },
        "WorkingItem": {
            "type": "object",
            "description": "Details for a working item",
            "properties": {
                "label": {
                    "type": "string"
                },
                "id": {
                    "type": "number",
                    "format": "integer"
                },
                "state": {
                    "type": "string"
                },
                "attributeConfiguration": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/AttributeConfiguration"
                    }
                }
            }
        },
        "AttributeConfiguration": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "id": {
                    "type": "number",
                    "format": "integer"
                },
                "type": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
        "Error": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "description": "Description of the error."
                }
            }
        },
...
    }
}